+++ meerschaum/api/tables/__init__.py
@@ -7,19 +7,7 @@ Create tables using the API instance connector
 """
 
 def get_tables(debug : bool = False, **kw):
-    """Call the normal get_tables with the API instance connector instead
-
-    Parameters
-    ----------
-    debug : bool :
-         (Default value = False)
-    **kw :
-        
-
-    Returns
-    -------
-
-    """
+    """Call the normal get_tables with the API instance connector instead."""
     from meerschaum.connectors.sql.tables import get_tables as _get_tables
     from meerschaum.api import get_api_connector
     ### get API instance connector