+++ meerschaum/connectors/sql/tables/__init__.py
@@ -114,7 +114,6 @@ def get_tables(
                 sqlalchemy.Column('user_type', sqlalchemy.String(256)),
                 sqlalchemy.Column('attributes', params_type),
                 extend_existing = True,
-                schema = conn.schema,
             ),
             'plugins' : sqlalchemy.Table(
                 *([
@@ -134,7 +133,6 @@ def get_tables(
                     sqlalchemy.ForeignKeyConstraint(['user_id'], ['users.user_id']),
                 ] if conn.flavor != 'duckdb' else [])),
                 extend_existing = True,
-                schema = conn.schema,
             ),
         }
 
@@ -154,7 +152,6 @@ def get_tables(
             sqlalchemy.Column("location_key", sqlalchemy.String(256), index=index_names),
             sqlalchemy.Column("parameters", params_type),
             extend_existing = True,
-            schema = conn.schema,
         )
 
         ### store the table dict for reuse (per connector)