+++ meerschaum/utils/dtypes/sql.py
@@ -7,18 +7,26 @@ Utility functions for working with SQL data types.
 """
 
 from __future__ import annotations
-from meerschaum.utils.typing import Dict, Union
+from meerschaum.utils.typing import Dict, Union, Tuple
+
+NUMERIC_PRECISION_FLAVORS: Dict[str, Tuple[int, int]] = {
+    'mariadb': (20, 10),
+    'mysql': (20, 10),
+    'mssql': (20, 10),
+    'duckdb': (20, 5),
+    'sqlite': (20, 7),
+}
 
 DB_TO_PD_DTYPES: Dict[str, Union[str, Dict[str, str]]] = {
     'FLOAT': 'float64[pyarrow]',
     'DOUBLE_PRECISION': 'float64[pyarrow]',
     'DOUBLE': 'float64[pyarrow]',
-    'DECIMAL': 'object',
+    'DECIMAL': 'numeric',
     'BIGINT': 'int64[pyarrow]',
     'INT': 'int64[pyarrow]',
     'INTEGER': 'int64[pyarrow]',
-    'NUMBER': 'float64[pyarrow]',
-    'NUMERIC': 'object',
+    'NUMBER': 'numeric',
+    'NUMERIC': 'numeric',
     'TIMESTAMP': 'datetime64[ns]',
     'TIMESTAMP WITH TIMEZONE': 'datetime64[ns, UTC]',
     'TIMESTAMPTZ': 'datetime64[ns, UTC]',
@@ -33,18 +41,20 @@ DB_TO_PD_DTYPES: Dict[str, Union[str, Dict[str, str]]] = {
     'TINYINT(1)': 'bool[pyarrow]',
     'BIT': 'bool[pyarrow]',
     'BIT(1)': 'bool[pyarrow]',
-    'JSON': 'object',
-    'JSONB': 'object',
+    'JSON': 'json',
+    'JSONB': 'json',
     'substrings': {
         'CHAR': 'string[pyarrow]',
         'TIMESTAMP': 'datetime64[ns]',
         'TIME': 'datetime64[ns]',
         'DATE': 'datetime64[ns]',
         'DOUBLE': 'double[pyarrow]',
-        'DECIMAL': 'object',
+        'DECIMAL': 'numeric',
+        'NUMERIC': 'numeric',
+        'NUMBER': 'numeric',
         'INT': 'int64[pyarrow]',
         'BOOL': 'bool[pyarrow]',
-        'JSON': 'object',
+        'JSON': 'json',
     },
     'default': 'object',
 }
@@ -122,7 +132,7 @@ PD_TO_DB_DTYPES_FLAVORS: Dict[str, Dict[str, str]] = {
         'mysql': 'BOOLEAN',
         'mssql': 'INTEGER',
         'oracle': 'INTEGER',
-        'sqlite': 'BOOLEAN',
+        'sqlite': 'INTEGER',
         'duckdb': 'BOOLEAN',
         'citus': 'BOOLEAN',
         'cockroachdb': 'BOOLEAN',
@@ -170,11 +180,11 @@ PD_TO_DB_DTYPES_FLAVORS: Dict[str, Dict[str, str]] = {
     'numeric': {
         'timescaledb': 'NUMERIC',
         'postgresql': 'NUMERIC',
-        'mariadb': 'NUMERIC',
-        'mysql': 'NUMERIC',
-        'mssql': 'NUMERIC',
-        'oracle': 'NUMERIC',
-        'sqlite': 'NUMERIC',
+        'mariadb': f'DECIMAL{NUMERIC_PRECISION_FLAVORS["mariadb"]}',
+        'mysql': f'DECIMAL{NUMERIC_PRECISION_FLAVORS["mysql"]}',
+        'mssql': f'NUMERIC{NUMERIC_PRECISION_FLAVORS["mssql"]}',
+        'oracle': 'NUMBER',
+        'sqlite': f'DECIMAL{NUMERIC_PRECISION_FLAVORS["sqlite"]}',
         'duckdb': 'NUMERIC',
         'citus': 'NUMERIC',
         'cockroachdb': 'NUMERIC',
@@ -240,8 +250,8 @@ PD_TO_SQLALCHEMY_DTYPES_FLAVORS: Dict[str, Dict[str, str]] = {
         'mariadb': 'Boolean',
         'mysql': 'Boolean',
         'mssql': 'Boolean',
-        'oracle': 'Boolean',
-        'sqlite': 'Boolean',
+        'oracle': 'Integer',
+        'sqlite': 'Integer',
         'duckdb': 'Boolean',
         'citus': 'Boolean',
         'cockroachdb': 'Boolean',
@@ -419,4 +429,10 @@ def get_db_type_from_pd_type(
     if db_type == 'JSONB':
         sqlalchemy_dialects_postgresql = attempt_import('sqlalchemy.dialects.postgresql')
         return sqlalchemy_dialects_postgresql.JSONB
+    if 'numeric' in db_type.lower():
+        numeric_type_str = PD_TO_DB_DTYPES_FLAVORS['numeric'].get(flavor, 'NUMERIC')
+        if flavor not in NUMERIC_PRECISION_FLAVORS:
+            return sqlalchemy_types.Numeric
+        precision, scale = NUMERIC_PRECISION_FLAVORS[flavor]
+        return sqlalchemy_types.Numeric(precision, scale)
     return getattr(sqlalchemy_types, db_type)